+++ sdk/communication/azure-communication-phonenumbers/azure/communication/phonenumbers/_phone_numbers_client.py
@@ -44,7 +44,7 @@ from ._shared.utils import parse_connection_str
 from ._version import SDK_MONIKER
 from ._api_versions import DEFAULT_VERSION
 
-_DEFAULT_POLLING_INTERVAL_IN_SECONDS = 2
+_DEFAULT_POLLING_INTERVAL_IN_SECONDS = 2  # Default polling interval for long-running operations
 
 
 class PhoneNumbersClient:
@@ -57,9 +57,12 @@ class PhoneNumbersClient:
     :param Union[TokenCredential, AzureKeyCredential] credential:
         The credential we use to authenticate against the service.
     :keyword api_version: Azure Communication Phone Number API version.
-        The default value is "2022-01-11-preview2".
+        The default value is "2025-06-01".
         Note that overriding this default value may result in unsupported behavior.
     :paramtype api_version: str
+    :keyword accepted_language: The locale to display in the localized fields in responses. e.g.
+        'en-US'. Default value is None.
+    :paramtype accepted_language: str
     """
 
     def __init__(self, endpoint: str, credential: Union[TokenCredential, AzureKeyCredential], **kwargs: Any) -> None:
@@ -259,7 +262,7 @@ class PhoneNumbersClient:
         return self._phone_number_client.phone_numbers.get_by_number(phone_number, **kwargs)
 
     @distributed_trace
-    def list_purchased_phone_numbers(self, **kwargs: Any) -> ItemPaged[PurchasedPhoneNumber]:
+    def list_purchased_phone_numbers(self, *, skip: int = 0, **kwargs: Any) -> ItemPaged[PurchasedPhoneNumber]:
         """Gets the list of all purchased phone numbers.
 
         :keyword skip: An optional parameter for how many entries to skip, for pagination purposes. The
@@ -271,10 +274,10 @@ class PhoneNumbersClient:
         :returns: An iterator of purchased phone numbers.
         :rtype: ~azure.core.paging.ItemPaged[~azure.communication.phonenumbers.PurchasedPhoneNumber]
         """
-        return self._phone_number_client.phone_numbers.list_phone_numbers(**kwargs)
+        return self._phone_number_client.phone_numbers.list_phone_numbers(skip=skip, **kwargs)
 
     @distributed_trace
-    def list_available_countries(self, **kwargs: Any) -> ItemPaged[PhoneNumberCountry]:
+    def list_available_countries(self, *, skip: int = 0, **kwargs: Any) -> ItemPaged[PhoneNumberCountry]:
         """Gets the list of supported countries.
 
         Gets the list of supported countries.
@@ -283,22 +286,32 @@ class PhoneNumbersClient:
          default value is 0. Default value is 0.
         :paramtype skip: int
         :return: An iterator like instance of PhoneNumberCountry
-        :rtype:
-         ~azure.core.paging.ItemPaged[~azure.communication.phonenumbers.PhoneNumberCountry]
+        :rtype: ~azure.core.paging.ItemPaged[~azure.communication.phonenumbers.PhoneNumberCountry]
         :raises ~azure.core.exceptions.HttpResponseError:
         """
         return self._phone_number_client.phone_numbers.list_available_countries(
-            accept_language=self._accepted_language, **kwargs
+            skip=skip, accept_language=self._accepted_language, **kwargs
         )
 
     @distributed_trace
-    def list_available_localities(self, country_code: str, **kwargs: Any) -> ItemPaged[PhoneNumberLocality]:
+    def list_available_localities(
+        self,
+        country_code: str,
+        *,
+        phone_number_type: Optional[Union[PhoneNumberType, str]] = None,
+        administrative_division: Optional[str] = None,
+        skip: int = 0,
+        **kwargs: Any
+    ) -> ItemPaged[PhoneNumberLocality]:
         """Gets the list of cities or towns with available phone numbers.
 
         Gets the list of cities or towns with available phone numbers.
 
         :param country_code: The ISO 3166-2 country/region two letter code, e.g. US. Required.
         :type country_code: str
+        :keyword phone_number_type: An optional parameter for the type of phone numbers, 
+         e.g. geographic, tollFree, mobile. Default value is None.
+        :paramtype phone_number_type: str or ~azure.communication.phonenumbers.PhoneNumberType
         :keyword administrative_division: An optional parameter for the name of the state or province
          in which to search for the area code. e.g. California. Default value is None.
         :paramtype administrative_division: str
@@ -306,61 +319,78 @@ class PhoneNumbersClient:
          default value is 0. Default value is 0.
         :paramtype skip: int
         :return: An iterator like instance of PhoneNumberLocality
-        :rtype:
-         ~azure.core.paging.ItemPaged[~azure.communication.phonenumbers.PhoneNumberLocality]
+        :rtype: ~azure.core.paging.ItemPaged[~azure.communication.phonenumbers.PhoneNumberLocality]
         :raises ~azure.core.exceptions.HttpResponseError:
         """
         return self._phone_number_client.phone_numbers.list_available_localities(
             country_code,
-            administrative_division=kwargs.pop(
-                "administrative_division", None),
+            administrative_division=administrative_division,
             accept_language=self._accepted_language,
+            phone_number_type=phone_number_type,
+            skip=skip,
             **kwargs
         )
 
     @distributed_trace
-    def list_available_offerings(self, country_code: str, **kwargs: Any) -> ItemPaged[PhoneNumberOffering]:
+    def list_available_offerings(
+        self,
+        country_code: str,
+        *,
+        phone_number_type: Optional[Union[PhoneNumberType, str]] = None,
+        assignment_type: Optional[Union[PhoneNumberAssignmentType, str]] = None,
+        skip: int = 0,
+        **kwargs: Any
+    ) -> ItemPaged[PhoneNumberOffering]:
         """List available offerings of capabilities with rates for the given country/region.
 
         List available offerings of capabilities with rates for the given country/region.
 
         :param country_code: The ISO 3166-2 country/region two letter code, e.g. US. Required.
         :type country_code: str
-        :keyword phone_number_type: Filter by phoneNumberType, e.g. Geographic, TollFree. Known values
-         are: "geographic" and "tollFree". Default value is None.
-        :paramtype phone_number_type: ~azure.communication.phonenumbers.PhoneNumberType
-        :keyword assignment_type: Filter by assignmentType, e.g. User, Application. Known values are:
+        :keyword phone_number_type: Filter by phone number type, e.g. geographic, tollFree, mobile.
+         Default value is None.
+        :paramtype phone_number_type: str or ~azure.communication.phonenumbers.PhoneNumberType
+        :keyword assignment_type: Filter by assignment type, e.g. User, Application. Known values are:
          "person" and "application". Default value is None.
-        :paramtype assignment_type: ~azure.communication.phonenumbers.PhoneNumberAssignmentType
+        :paramtype assignment_type: str or ~azure.communication.phonenumbers.PhoneNumberAssignmentType
         :keyword skip: An optional parameter for how many entries to skip, for pagination purposes. The
          default value is 0. Default value is 0.
         :paramtype skip: int
         :return: An iterator like instance of PhoneNumberOffering
-        :rtype:
-         ~azure.core.paging.ItemPaged[~azure.communication.phonenumbers.PhoneNumberOffering]
+        :rtype: ~azure.core.paging.ItemPaged[~azure.communication.phonenumbers.PhoneNumberOffering]
         :raises ~azure.core.exceptions.HttpResponseError:
         """
         return self._phone_number_client.phone_numbers.list_offerings(
             country_code,
-            phone_number_type=kwargs.pop("phone_number_type", None),
-            assignment_type=kwargs.pop("assignment_type", None),
+            phone_number_type=phone_number_type,
+            assignment_type=assignment_type,
+            accept_language=self._accepted_language,
+            skip=skip,
             **kwargs
         )
 
     @distributed_trace
     def list_available_area_codes(
-        self, country_code: str, phone_number_type: PhoneNumberType, **kwargs: Any
+        self,
+        country_code: str,
+        phone_number_type: Union[PhoneNumberType, str],
+        *,
+        administrative_division: Optional[str] = None,
+        assignment_type: Optional[Union[PhoneNumberAssignmentType, str]] = None,
+        locality: Optional[str] = None,
+        skip: int = 0,
+        **kwargs: Any
     ) -> ItemPaged[PhoneNumberAreaCode]:
         """Gets the list of available area codes.
 
         :param country_code: The ISO 3166-2 country/region two letter code, e.g. US. Required.
         :type country_code: str
-        :param phone_number_type: Filter by phone number type, e.g. Geographic, TollFree. Known values are:
-         "geographic" and "tollFree". Required.
-        :type phone_number_type: ~azure.communication.phonenumbers.PhoneNumberType
-        :keyword assignment_type: Filter by assignmentType, e.g. User, Application. Known values are:
+        :param phone_number_type: Filter by phone number type, e.g. geographic, tollFree, mobile.
+         Required.
+        :type phone_number_type: str or ~azure.communication.phonenumbers.PhoneNumberType
+        :keyword assignment_type: Filter by assignment type, e.g. User, Application. Known values are:
          "person" and "application". Default value is None.
-        :paramtype assignment_type: ~azure.communication.phonenumbers.PhoneNumberAssignmentType
+        :paramtype assignment_type: str or ~azure.communication.phonenumbers.PhoneNumberAssignmentType
         :keyword locality: The name of locality in which to search for the area code. e.g. Seattle.
          This is required if the phone number type is Geographic. Default value is None.
         :paramtype locality: str
@@ -377,16 +407,21 @@ class PhoneNumbersClient:
         return self._phone_number_client.phone_numbers.list_area_codes(
             country_code,
             phone_number_type=phone_number_type,
-            assignment_type=kwargs.pop("assignment_type", None),
-            locality=kwargs.pop("locality", None),
-            administrative_division=kwargs.pop(
-                "administrative_division", None),
+            accept_language=self._accepted_language,
+            assignment_type=assignment_type,
+            locality=locality,
+            administrative_division=administrative_division,
+            skip=skip,
             **kwargs
         )
 
     @distributed_trace
     def search_operator_information(
-        self, phone_numbers: Union[str, List[str]], *, options: Optional[OperatorInformationOptions] = None, **kwargs: Any  # pylint: disable=line-too-long
+        self,
+        phone_numbers: Union[str, List[str]],
+        *,
+        options: Optional[OperatorInformationOptions] = None,
+        **kwargs: Any
     ) -> OperatorInformationResult:
         """Searches for operator information for a given list of phone numbers.
 
@@ -436,8 +471,7 @@ class PhoneNumbersClient:
          purposes. The default value is 100. Default value is 100.
         :paramtype max_page_size: int
         :return: An iterator like instance of PhoneNumbersReservation
-        :rtype:
-         ~azure.core.paging.ItemPaged[~azure.communication.phonenumbers.PhoneNumbersReservation]
+        :rtype: ~azure.core.paging.ItemPaged[~azure.communication.phonenumbers.PhoneNumbersReservation]
         """
 
         # This allows mapping the generated model to the public model.
@@ -449,7 +483,8 @@ class PhoneNumbersClient:
 
     @distributed_trace
     def create_or_update_reservation(
-        self, *,
+        self,
+        *,
         reservation_id: str,
         numbers_to_add: Optional[List[AvailablePhoneNumber]] = None,
         numbers_to_remove: Optional[List[str]] = None,