+++ sdk/evaluation/azure-ai-evaluation/azure/ai/evaluation/_evaluators/_document_retrieval/__init__.py
@@ -4,8 +4,4 @@
 
 from ._document_retrieval import DocumentRetrievalEvaluator, RetrievalGroundTruthDocument, RetrievedDocument
 
-__all__ = [
-    "DocumentRetrievalEvaluator",
-    "RetrievalGroundTruthDocument",
-    "RetrievedDocument"
-]
+__all__ = ["DocumentRetrievalEvaluator", "RetrievalGroundTruthDocument", "RetrievedDocument"]