+++ sdk/identity/azure-identity/tests/test_context_manager.py
@@ -24,7 +24,6 @@ from azure.identity._constants import EnvironmentVariables
 import pytest
 
 from test_certificate_credential import PEM_CERT_PATH
-from test_vscode_credential import GET_USER_SETTINGS
 
 
 class CredentialFixture:
@@ -62,7 +61,6 @@ FIXTURES = (
         {kwarg: "..." for kwarg in ("tenant_id", "client_id", "client_secret", "user_assertion")},
     ),
     CredentialFixture(UsernamePasswordCredential, {"client_id": "...", "username": "...", "password": "..."}),
-    CredentialFixture(VisualStudioCodeCredential, ctor_patch_factory=lambda: patch(GET_USER_SETTINGS, lambda: {})),
 )
 
 all_fixtures = pytest.mark.parametrize("fixture", FIXTURES, ids=lambda fixture: fixture.cls.__name__)