+++ sdk/storage/azure-mgmt-storage/azure/mgmt/storage/aio/operations/_network_security_perimeter_configurations_operations.py
@@ -6,7 +6,7 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 from collections.abc import MutableMapping
-from typing import Any, AsyncIterable, AsyncIterator, Callable, Dict, Optional, TypeVar, Union, cast
+from typing import Any, AsyncIterator, Callable, Dict, Optional, TypeVar, Union, cast
 import urllib.parse
 
 from azure.core import AsyncPipelineClient
@@ -65,7 +65,7 @@ class NetworkSecurityPerimeterConfigurationsOperations:  # pylint: disable=name-
     @distributed_trace
     def list(
         self, resource_group_name: str, account_name: str, **kwargs: Any
-    ) -> AsyncIterable["_models.NetworkSecurityPerimeterConfiguration"]:
+    ) -> AsyncItemPaged["_models.NetworkSecurityPerimeterConfiguration"]:
         """Gets list of effective NetworkSecurityPerimeterConfiguration for storage account.
 
         :param resource_group_name: The name of the resource group within the user's subscription. The